Refactor environment and knowledge graph initialization to include converter parameter for enhanced feature embedding support. Now empty entity nodes are not automaticcaly masked.

+++ src/tsp_rl_kg/rl/agent_model.py
@@ -6,7 +6,7 @@ from torch_geometric.data import Data
 from stable_baselines3.common.torch_layers import BaseFeaturesExtractor
 import gymnasium as gym
 
-torch_dtype = torch.float32
+torch_dtype = torch.float64
 
 class VisionProcessor(BaseFeaturesExtractor):
     """
